+++ tests/constructs/test_lambdas.py
@@ -15,13 +15,13 @@ from aws_cdk import (
     core
 )
 
-from aws_emr_launch.constructs.lambdas import EMRLambdas
+from aws_emr_launch.constructs.lambdas.emr_control import EMRControlLambdas
 
 
 def test_emr_lambdas():
     app = core.App()
     stack = core.Stack(app, 'test-stack')
-    emr_lambdas = EMRLambdas(stack, 'test-lambdas')
+    emr_lambdas = EMRControlLambdas(stack, 'test-lambdas')
 
     assert emr_lambdas.run_job_flow
     assert emr_lambdas.add_job_flow_steps