+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog for emr-launch
 =============================
 
-0.1.1 (unreleased)
+0.1.1 (2019-11-01)
 ------------------
 
 - Store EMR Cluster Configs